QDSP6v2 Codecs Integrated - Illustrious http://cafe.naver.com/skydevelopers/181462

+++ arch/arm/mach-msm/qdsp6v2/pcm_in.c
@@ -114,7 +114,7 @@ static int pcm_in_enable(struct pcm *pcm)
 	if (atomic_read(&pcm->in_enabled))
 		return 0;
 #ifdef CONFIG_SKYSND_CTRL
-              pcm_in_status=1;
+              pcm_in_status = 1;
 #endif
 	return q6asm_run(pcm->ac, 0, 0, 0);
 }
@@ -124,8 +124,8 @@ static int pcm_in_disable(struct pcm *pcm)
 	int rc = 0;
 
 #ifdef CONFIG_SKYSND_CTRL
-              pcm_in_status=0;
-              pcm_mute_exception=0;
+              pcm_in_status = 0;
+              pcm_mute_exception = 0;
 #endif
 	if (atomic_read(&pcm->in_opened)) {
 		atomic_set(&pcm->in_enabled, 0);
@@ -423,11 +423,10 @@ static ssize_t pcm_in_read(struct file *file, char __user *buf,
 		if ((len) && data) {
 			offset = pcm->in_frame_info[idx][1];
                      #ifdef CONFIG_SKYSND_CTRL  // SangwonLee 110406 To remove camera effect sound.
-                     pcm_offset=(int)(48000/pcm->sample_rate);
-                     if(pcm_mute_exception>=pcm_offset) {
+                     pcm_offset = 48000 / pcm->sample_rate;
+                     if(pcm_mute_exception >= pcm_offset) {
                         memset(data+offset, 0, size);
-                        //pr_info("%s:Erase Tx data (size:%d, seq:%d, rate:%d)\n", __func__, size, pcm_mute_exception, pcm->sample_rate);
-                        pcm_mute_exception-=pcm_offset;
+                        pcm_mute_exception -= pcm_offset;
                      }
                      #endif
 			if (copy_to_user(buf, data+offset, len)) {
@@ -514,12 +513,12 @@ int get_pcm_in_status(void)
 #if 1//SangwonLee 110330 Right Speaker is near by handset mic. Disable it to prevent recording noise.
 void set_pcm_in_status(void)
 {
-	pcm_in_status=1;
+	pcm_in_status = 1;
 }
 #endif
 void set_pcm_in_exception(int value)
 {
-    pcm_mute_exception=value;
+    pcm_mute_exception = value;
 }
 #endif
 device_initcall(pcm_in_init);
